[Loc](12:31PM): Handle api loader; phone,tablet loader, product detail loader, cart loader
