+++ src/apis/apiSlice.js
@@ -10,8 +10,8 @@ import {
 const baseQuery = fetchBaseQuery({
   // local
   // baseUrl: "https://tgddgroup04.herokuapp.com/api",
-    baseUrl: "https://tlcn-2022-be.onrender.com/api",
-  // baseUrl: "http://localhost:5000/api",
+    // baseUrl: "https://tlcn-2022-be.onrender.com/api",
+  baseUrl: "http://localhost:5000/api",
   // baseUrl: "http://192.168.1.6:5000/api",
 
   credentials: "include",
