+++ src/components/cart/subComponent/Info.jsx
@@ -84,8 +84,7 @@ const Info = ({ orderInfo, setOrderInfo, setDetailAddress, detailAddress }) => {
           const defaultAddress =
             user?.addresses.find((v) => v.idDefault === true) ||
             user?.addresses?.[0];
-          console.clear();
-          console.log(user?.addresses);
+
           setAddressEdit(defaultAddress?.detailAddress);
           setCurrentUser({
             ...currentUser,
