+++ src/components/phone/dummyData.js
@@ -1,0 +1,36 @@
+const manu = [{
+        id: "0",
+        name: "Apple",
+    },
+    {
+        id: "1",
+        name: "Samsung",
+    },
+];
+const price = [{
+        id: "0",
+        name: "Dưới 5 triệu",
+    },
+    {
+        id: "1",
+        name: "Từ 5 triệu đến 10 triệu",
+    },
+    {
+        id: "2",
+        name: "Từ 10 đến 20 triệu",
+    },
+    {
+        id: "3",
+        name: "Trên 20 triệu",
+    },
+];
+
+const os = [{
+        id: "0",
+        name: "IOS",
+    },
+    {
+        id: "1",
+        name: "Android",
+    },
+];
