+++ src/pages/ProductDetail.jsx
@@ -11,6 +11,7 @@ import productHandler from "../features/product/function";
 import StarRating from "../components/accessories/StarRating";
 import "../sass/phone/rating.scss";
 import "../sass/productdetail/_product_detail.scss";
+import Loader from "../components/loading/Loader";
 const ProductDetail = () => {
   const location = useLocation();
 
@@ -21,10 +22,10 @@ const ProductDetail = () => {
   //console.log(location);
 
   const [product, setProduct] = useState();
-
+  const [isLoading, setIsLoading] = useState(true);
   useEffect(() => {
     productHandler.getProductById({ productId: productId }).then((res) => {
-      console.log(res);
+      setIsLoading(res.isLoading);
       const product = res.data;
       // console.log("Get product by Id:");
       // console.log(product);
@@ -32,64 +33,53 @@ const ProductDetail = () => {
     });
   }, [productId]);
 
-  // useEffect(() => {
-  //   productHandler.getProductById({ productId: productId }).then((res) => {
-  //     // console.log(res);
-  //     const product = res.data;
-  //     // console.log("Get product Test:");
-  //     // console.log(product);
-  //     // console.log(product?.category);
-  //     // console.log(product?.manufacturer);
-  //     // console.log(product?.reviews?.lenght);
-  //     setProduct(product);
-  //   });
-  // }, []);
-
   useEffect(() => {
     window.scrollTo(0, 0);
   });
 
   return (
-    <div className="produt_detail grid  wide">
-      {/* <!-- category --> */}
-      <div className="product_category">
-        <span className="product_category-title ">{product?.category}</span>
-        <span className="product_category-title">
-          <i className="fa-solid fa-angle-right"></i>
-        </span>
-        <span className="product_category-title">
-          {product?.category} {product?.manufacturer}
-        </span>
-      </div>
-
-      {/* <!-- title --> */}
-      <div className="product_title">
-        <h2 className="product_title_name">{product?.name}</h2>
-        <div className="product_title_star">
-          <div className="star-phone">
-            <StarRating rating={product?.rating} />
-          </div>
+    <Loader isLoading={isLoading}>
+      <div className="produt_detail grid  wide">
+        {/* <!-- category --> */}
+        <div className="product_category">
+          <span className="product_category-title ">{product?.category}</span>
+          <span className="product_category-title">
+            <i className="fa-solid fa-angle-right"></i>
+          </span>
+          <span className="product_category-title">
+            {product?.category} {product?.manufacturer}
+          </span>
         </div>
-        {/* <div className="product_title_star">
+
+        {/* <!-- title --> */}
+        <div className="product_title">
+          <h2 className="product_title_name">{product?.name}</h2>
+          <div className="product_title_star">
+            <div className="star-phone">
+              <StarRating rating={product?.rating} />
+            </div>
+          </div>
+          {/* <div className="product_title_star">
           <i className="fa-solid fa-star"></i>
           <i className="fa-solid fa-star"></i>
           <i className="fa-solid fa-star"></i>
           <i className="fa-solid fa-star"></i>
           <i className="fa-solid fa-star-half-stroke"></i>
         </div> */}
-        <span className="product_title_review flex_center">
-          {`${product?.reviews.length || "Không có"} đánh giá `}
-        </span>
+          <span className="product_title_review flex_center">
+            {`${product?.reviews.length || "Không có"} đánh giá `}
+          </span>
+        </div>
+        <div className="line"></div>
+        <Introduce product={product} />
+        <div className="line"></div>
+        <Information product={product} />
+        <div className="line"></div>
+        <Feedback product={product} />
+        <div className="line"></div>
+        <SeeMore />
       </div>
-      <div className="line"></div>
-      <Introduce product={product} />
-      <div className="line"></div>
-      <Information product={product} />
-      <div className="line"></div>
-      <Feedback product={product} />
-      <div className="line"></div>
-      <SeeMore />
-    </div>
+    </Loader>
   );
 };
